// packages/eorm-wizard/src/generated/AttributeConfig.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EStructuralFeature, BasicEObject, EClass } from '@emfts/core';
import type { AttributeRole } from './AttributeRole';
import type { IdStrategy } from './IdStrategy';

/**
 * AttributeConfig
 * @generated
 */
export interface AttributeConfig extends EObject {
  selected?: boolean;
  feature: EStructuralFeature;
  role: AttributeRole;
  columnName?: string;
  nullable?: boolean;
  unique?: boolean;
  length?: number;
  idStrategy?: IdStrategy;
  sequenceName?: string;

}