// packages/eorm-wizard/src/generated/EntityConfig.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EClass, BasicEObject, EStructuralFeature } from '@emfts/core';
import type { InheritanceStrategy } from './InheritanceStrategy';
import type { AttributeConfig } from './AttributeConfig';
import type { RelationConfig } from './RelationConfig';

/**
 * EntityConfig
 * @generated
 */
export interface EntityConfig extends EObject {
  selected?: boolean;
  targetClass: EClass;
  entityName?: string;
  tableName?: string;
  inheritance?: InheritanceStrategy;
  discriminatorColumn?: string;
  attributes: AttributeConfig[];
  relations: RelationConfig[];

}