// packages/eorm-wizard/src/generated/RelationConfig.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EReference, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { RelationKind } from './RelationKind';
import type { JoinStrategy } from './JoinStrategy';
import type { FetchMode } from './FetchMode';

/**
 * RelationConfig
 * @generated
 */
export interface RelationConfig extends EObject {
  selected?: boolean;
  feature: EReference;
  kind: RelationKind;
  joinStrategy: JoinStrategy;
  joinName?: string;
  fetch: FetchMode;
  cascadeAll?: boolean;
  orphanRemoval?: boolean;
  mappedBy?: string;
  batch?: boolean;

}