// packages/eorm-wizard/src/generated/IdStrategy.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * IdStrategy
 * @generated
 */
export const IdStrategy = {
  NONE: 'NONE',
  SEQUENCE: 'SEQUENCE',
  UUID: 'UUID',
  IDENTITY: 'IDENTITY',
  TABLE: 'TABLE'
} as const;

export type IdStrategy = typeof IdStrategy[keyof typeof IdStrategy];

/**
 * Get enum value by name
 */
export function getIdStrategy(name: string): IdStrategy | undefined {
  return (IdStrategy as Record<string, string>)[name] as IdStrategy | undefined;
}

/**
 * Get enum value by literal
 */
export function getIdStrategyByLiteral(literal: string): IdStrategy | undefined {
  if (literal === 'NONE') return IdStrategy.NONE;
  if (literal === 'SEQUENCE') return IdStrategy.SEQUENCE;
  if (literal === 'UUID') return IdStrategy.UUID;
  if (literal === 'IDENTITY') return IdStrategy.IDENTITY;
  if (literal === 'TABLE') return IdStrategy.TABLE;
  return undefined;
}